Delete button at course members page; style home page

+++ pages/user/[userId]/course/[courseId]/member/index.js
@@ -5,7 +5,7 @@ import clsx from 'clsx';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 import { isObject } from 'lodash';
-import { useState, useEffect, useMemo, useContext } from 'react';
+import { useState, useEffect, useMemo, useContext, useCallback } from 'react';
 
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
@@ -26,6 +26,7 @@ import withLayout from '../../../../../../components/hoc/withLayout';
 import withCourseLayout from '../../../../../../components/hoc/withCourseLayout';
 import AbsURL from '../../../../../../components/helpers/URL';
 import AppUser from '../../../../../../components/auth/AppUser';
+import { MemberType, CourseType } from '../../../../../../components/propTypes';
 
 const useStyles = makeStyles((theme) => ({
   memberContainer: {
@@ -40,9 +41,16 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-function MemberItem(props) {
+const MemberItem = (props) => {
   const classes = useStyles();
   const { member, onDelete } = props;
+  const userContext = useContext(AppUser);
+
+  const isCourseOwner = useMemo(() => {
+    if (isObject(userContext.user)) {
+      return userContext.user.id === member.teacher_id;
+    }
+  }, [userContext.user, member.teacher_id]);
   return (
     <Grid item xs={12}>
       <Paper className={clsx(classes.memberContainer)}>
@@ -61,16 +69,18 @@ function MemberItem(props) {
               <Typography variant="caption">{dayjs(member.joined_at).format('YYYY MMM D hh:mm A')}</Typography>
             </Box>
           </Grid>
-          <Grid item>
-            <IconButton onClick={onDelete}>
-              <Icon>delete</Icon>
-            </IconButton>
-          </Grid>
+          {isCourseOwner && (
+            <Grid item>
+              <IconButton onClick={onDelete}>
+                <Icon>delete</Icon>
+              </IconButton>
+            </Grid>
+          )}
         </Grid>
       </Paper>
     </Grid>
   );
-}
+};
 
 function CourseMember(props) {
   const { userId, courseId, memberData, course, page } = props;
@@ -78,28 +88,29 @@ function CourseMember(props) {
   const [currentPage, setPage] = useState(page);
   const [members, setMembers] = useState(memberData.members);
   const router = useRouter();
-  const userCotext = useContext(AppUser);
 
-  const handleDelete = async (event, id) => {
+  const handleDelete = useCallback(async (event, member) => {
     event.preventDefault();
-    setMembers((members) => members.filter((member) => member.student_id !== id));
-    try {
-      const resFetch = await fetch(AbsURL(`/api/user/${userId}/course/${courseId}/member/${id}`), {
-        method: 'DELETE',
-        credentials: 'include',
-        mode: 'same-origin'
-      });
-      const res = await resFetch.json();
-      if (res.successful) {
-        console.log('Delete successfully');
-        return;
-      } else {
-        console.log(res);
+    if (confirm(`Are you sure that you want to remove "${member.username}" from this course?`)) {
+      setMembers((members) => members.filter((currentMember) => currentMember.student_id !== member.student_id));
+      try {
+        const resFetch = await fetch(AbsURL(`/api/user/${userId}/course/${courseId}/member/${member.student_id}`), {
+          method: 'DELETE',
+          credentials: 'include',
+          mode: 'same-origin'
+        });
+        const res = await resFetch.json();
+        if (res.successful) {
+          console.log('Delete successfully');
+          return;
+        } else {
+          console.log(res);
+        }
+      } catch (error) {
+        console.error(error);
       }
-    } catch (error) {
-      console.error(error);
     }
-  };
+  }, []);
 
   useEffect(() => {
     router.push(
@@ -108,15 +119,10 @@ function CourseMember(props) {
     );
   }, [currentPage]);
 
-  const isCourseOwner = useMemo(() => {
-    if (isObject(userCotext.user)) {
-      return userCotext.user.id === userId;
-    }
-  }, [userCotext.user, userId]);
   return (
     <>
       <Head>
-        <title>{`${course.course_name} 's member`}</title>
+        <title>{`${course.course_name}'s member`}</title>
       </Head>
       <Box py={2} />
       {memberData.total ? (
@@ -126,7 +132,7 @@ function CourseMember(props) {
               return (
                 <TableRow key={current.student_id}>
                   <TableCell className={clsx(classes.cell)}>
-                    <MemberItem member={current} onDelete={(event) => handleDelete(event, current.student_id)} />
+                    <MemberItem member={current} onDelete={(event) => handleDelete(event, current)} />
                   </TableCell>
                 </TableRow>
               );
@@ -177,16 +183,19 @@ CourseMember.getInitialProps = async (context) => {
 };
 
 MemberItem.propTypes = {
-  member: PropTypes.object.isRequired,
+  member: MemberType.isRequired,
   onDelete: PropTypes.func.isRequired
 };
 
 CourseMember.propTypes = {
   page: PropTypes.number.isRequired,
-  memberData: PropTypes.object.isRequired,
+  memberData: PropTypes.shape({
+    members: PropTypes.arrayOf(MemberType),
+    total: PropTypes.number
+  }),
   userId: PropTypes.string.isRequired,
   courseId: PropTypes.string.isRequired,
-  course: PropTypes.object.isRequired,
+  course: CourseType.isRequired,
   course_name: PropTypes.string
 };
